hydra-crystal-notify: Include jobset in eval status
I'm going to split the hydra jobset for cardano-wallet, to reduce
memory consumption of a single eval. I will probably also do it for
cardano-node.

But I realised that the GitHub CI status from hydra is
`ci/hydra-eval`, which would be ambiguous with multiple jobsets.

What would be great is to have the context name sent as
`ci/hydra-eval:Project:jobset`.

After this is deployed, any GitHub branch protections or `bors.toml`
files which have `ci/hydra-eval` would need to be updated.